% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_agent_analytics}
\alias{get_agent_analytics}
\title{Get agent analytics}
\usage{
get_agent_analytics(
  map_id = NULL,
  region_id = NULL,
  event_id = NULL,
  role_id = NULL,
  patch_id = NULL
)
}
\value{
a data.frame
}
\description{
Get agent analytics
}
\examples{
\donttest{
agent_analytics <- get_agent_analytics()
dplyr::glimpse(agent_analytics)
}
}
