% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_all_agent_names}
\alias{get_all_agent_names}
\title{Get all agents}
\usage{
get_all_agent_names()
}
\value{
a data.frame
}
\description{
Get all agent info
}
\examples{
all_agents <- get_all_agent_names()
dplyr::glimpse(all_agents)
}
