% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_all_armor_names}
\alias{get_all_armor_names}
\title{Get all armor}
\usage{
get_all_armor_names()
}
\value{
a data.frame with columns \code{armorId} and \code{armorName}
}
\description{
Get all armor info
}
\examples{
all_armors <- get_all_armor_names()
dplyr::glimpse(all_armors)
}
