% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_all_map_names}
\alias{get_all_map_names}
\title{Get all maps}
\usage{
get_all_map_names()
}
\value{
a data.frame with columns \code{mapId} and \code{mapName}
}
\description{
Get all map info
}
\examples{
all_maps <- get_all_map_names()
dplyr::glimpse(all_maps)
}
