% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_all_region_names}
\alias{get_all_region_names}
\title{Get all regions}
\usage{
get_all_region_names()
}
\value{
a data.frame with columns \code{regionId} and \code{regionName}
}
\description{
Get all region info
}
\examples{
all_regions <- get_all_region_names()
dplyr::glimpse(all_regions)
}
