% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_all_role_names}
\alias{get_all_role_names}
\title{Get all roles}
\usage{
get_all_role_names()
}
\value{
a data.frame
}
\description{
Get all role info
}
\examples{
all_roles <- get_all_role_names()
dplyr::glimpse(all_roles)
}
