% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_all_team_names}
\alias{get_all_team_names}
\title{Get all teams}
\usage{
get_all_team_names()
}
\value{
a data.frame
}
\description{
Get all team info
}
\examples{
\donttest{
all_teams <- get_all_team_names()
dplyr::glimpse(all_teams)
}
}
