% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_all_teams}
\alias{get_all_teams}
\title{Get all teams}
\usage{
get_all_teams()
}
\value{
a tibble
}
\description{
Get all teams' info
}
\examples{
\donttest{
all_teams <- get_all_teams()
dplyr::glimpse(all_teams)
}
}
