% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_composition_analytics}
\alias{get_composition_analytics}
\title{Get composition analytics}
\usage{
get_composition_analytics(
  map_id = NULL,
  region_id = NULL,
  event_id = NULL,
  role_id = NULL,
  patch_id = NULL
)
}
\value{
a data.frame
}
\description{
Get composition analytics
}
\examples{
\donttest{
composition_analytics <- get_composition_analytics()
dplyr::glimpse(composition_analytics)
}
}
