% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_events}
\alias{get_events}
\title{Get events}
\usage{
get_events(
  query,
  ...,
  sort_by = "startDate",
  ascending = FALSE,
  has_series = TRUE,
  n_results = 50
)
}
\arguments{
\item{query}{Text description of event, e.g. "VALORANT Champions 2022 Playoffs"}

\item{\dots}{Currently un-used}

\item{sort_by, ascending}{How to return the results}

\item{has_series}{Whether or not to return events with completed series}

\item{n_results}{Number of results to return. API tries to respect \code{n_results} even if event name doesn't match query. 50 is default of the API.}
}
\value{
a tibble
}
\description{
Get event ids given a free-text query
}
\details{
Note that sometimes the query doesn't seem to work. Results will be returned, but they are a default set of results.
}
\examples{
\donttest{
events <- get_events("VALORANT Champions 2022")
dplyr::glimpse(events)
}
}
