% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_match_details}
\alias{get_match_details}
\title{Get match details}
\usage{
get_match_details(match_id)
}
\arguments{
\item{match_id}{Match ID}
}
\value{
a list where each element represents a map. In each element is another list with the following elements: id, events, locations, economies.
}
\description{
Get match details for a given series
}
\examples{
\donttest{
match_details <- get_match_details(79018) ## Optic vs. Liquid, Map 3
match_details |> purrr::pluck("locations")
}
}
