% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_matches}
\alias{get_matches}
\title{Get matches}
\usage{
get_matches(series_id)
}
\arguments{
\item{series_id}{Series ID}
}
\value{
a list with the following elements
}
\description{
Get match ids for a given series
}
\examples{
\donttest{
matches <- get_matches(35225) ## Optic vs. Liquid
matches |> purrr::pluck("matches")
}
}
