% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_player}
\alias{get_player}
\title{Get player}
\usage{
get_player(player_id)
}
\arguments{
\item{player_id}{Player ID}
}
\value{
a list with the following elements
}
\description{
Get player info given a player ID
}
\examples{
\donttest{
yay <- get_player(2716)
dplyr::glimpse(yay)
}
}
