% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_series}
\alias{get_series}
\title{Get series}
\usage{
get_series(event_id, ..., completed = TRUE, n_results = 50)
}
\arguments{
\item{event_id}{Event ID}

\item{\dots}{Currently un-used}

\item{completed}{Whether or not to return completed series}

\item{n_results}{Number of results to return. 50 is default of the API.}
}
\value{
a tibble
}
\description{
Get series ids for a given event
}
\examples{
\donttest{
series <- get_series(1866) ## VALORANT Champions 2022 - Playoffs
dplyr::glimpse(series)
}
}
