% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_team}
\alias{get_team}
\title{Get team}
\usage{
get_team(team_id)
}
\arguments{
\item{team_id}{Team ID}
}
\value{
a data.frame
}
\description{
Get team info
}
\examples{
\donttest{
sentinels <- get_team(388)
str(sentinels, max.level = 1)
}
}
