% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_weapon_analytics}
\alias{get_weapon_analytics}
\title{Get weapon analytics}
\usage{
get_weapon_analytics(
  map_id = NULL,
  side = NULL,
  region_id = NULL,
  event_id = NULL,
  role_id = NULL,
  patch_id = NULL
)
}
\arguments{
\item{map_id}{Map ID (optional). See \code{get_all_map_names()}.}

\item{side}{If not \code{NULL}, then either \code{atk} or \code{def}.}

\item{region_id}{Region ID (optional). See \code{get_all_region_names()}.}

\item{event_id}{Event ID (optional). See \code{get_events()}.}

\item{role_id}{Role ID (optional). See \code{get_all_role_names()}.}

\item{patch_id}{Patch ID (optional). \code{patch_id = 12} is patch 4.0. \code{patch_id = 23} is patch 5.0. Versions are incremented by 1.}
}
\value{
a data.frame
}
\description{
Get weapon analytics
}
\examples{
\donttest{
weapon_analytics <- get_weapon_analytics()
dplyr::glimpse(weapon_analytics)
}
}
