% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_valorant}
\alias{load_valorant}
\title{Load a valorant-data release}
\usage{
load_valorant(tag, ...)
}
\arguments{
\item{tag}{Tag of release in valorant-data}

\item{\dots}{Extra params to pass to \code{piggyback::pb_releases} to check for existing releases.}
}
\description{
Load a valorant-data release
}
