% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{qs_from_url}
\alias{qs_from_url}
\title{Load a .qs file from a URL}
\source{
\url{https://github.com/nflverse/nflreadr/blob/main/R/from_url.R#L185}
}
\usage{
qs_from_url(url)
}
\arguments{
\item{url}{URL}
}
\description{
Load a .qs file from a URL
}
