% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_all_weapon_names}
\alias{get_all_weapon_names}
\title{Get all weapons}
\usage{
get_all_weapon_names()
}
\value{
a data.frame
}
\description{
Get all weapon info
}
\examples{
all_weapons <- get_all_weapon_names()
dplyr::glimpse(all_weapons)
}
