% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_map_analytics}
\alias{get_map_analytics}
\title{Get map analytics}
\usage{
get_map_analytics(region_id = NULL, event_id = NULL, patch_id = NULL)
}
\value{
a data.frame
}
\description{
Get map analytics
}
\examples{
\donttest{
map_analytics <- get_map_analytics()
dplyr::glimpse(map_analytics)
}
}
