% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_weapon_analytics}
\alias{get_weapon_analytics}
\title{Get weapon analytics}
\usage{
get_weapon_analytics()
}
\value{
a data.frame
}
\description{
Get weapon analytics
}
\examples{
\donttest{
weapon_analytics <- get_weapon_analytics()
dplyr::glimpse(weapon_analytics)
}
}
